+++ lib/login/login.dart
@@ -1,2 +1,2 @@
 export 'cubit/login_cubit.dart';
-export 'view/login_page.dart';
+export 'view/view.dart';
