+++ lib/login/view/login_page.dart
@@ -29,67 +29,3 @@ class LoginView extends StatelessWidget {
     );
   }
 }
-
-class LoginForm extends StatelessWidget {
-  const LoginForm({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        const Text(
-          'Login',
-          style: TextStyle(
-            fontSize: 32,
-            fontWeight: FontWeight.w600,
-          ),
-        ),
-        const SizedBox(height: 16),
-        const Text(
-          'Please enter your credentials',
-          style: TextStyle(
-            fontSize: 16,
-            fontWeight: FontWeight.w400,
-          ),
-        ),
-        const SizedBox(height: 16),
-        const Text(
-          'Username',
-          style: TextStyle(
-            fontSize: 16,
-            fontWeight: FontWeight.w400,
-          ),
-        ),
-        const SizedBox(height: 16),
-        const TextField(
-          decoration: InputDecoration(
-            border: OutlineInputBorder(),
-            labelText: 'Username',
-          ),
-        ),
-        const SizedBox(height: 16),
-        const Text(
-          'Password',
-          style: TextStyle(
-            fontSize: 16,
-            fontWeight: FontWeight.w400,
-          ),
-        ),
-        const SizedBox(height: 16),
-        const TextField(
-          decoration: InputDecoration(
-            border: OutlineInputBorder(),
-            labelText: 'Password',
-          ),
-        ),
-        const SizedBox(height: 16),
-        ElevatedButton(
-          onPressed: () {
-            // context.bloc<LoginCubit>().login();
-          },
-          child: const Text('Login'),
-        ),
-      ],
-    );
-  }
-}
