+++ lib/login/view/view.dart
@@ -1,1 +1,2 @@
+export 'login_form.dart';
 export 'login_page.dart';
